refactor(core): move audio analysis pipeline to core
Second slice of step 6 — `analysis.rs` (the per-track peak /
loudness / ReplayGain / BPM computer) moves verbatim to
`waveflow_core::analysis`. It had no Tauri / DB coupling, just
`symphonia` decode + plain f64 math, so the move is a `git mv` plus
a re-export stub in app for the single caller (`commands/analysis.rs`).

`crates/core/Cargo.toml` picks up the `symphonia` declaration with
the same codec feature set the desktop runtime needs — Cargo
unifies it back to a single sqlx-style compile across the workspace.

`cargo check --workspace --all-targets` clean; `cargo test
--workspace` 111 passed / 0 failed.

Refs #128.
